Update email escalations as follows
When the concession is 3 months due for expiry, the email notification is sent to the AE/AA.
When the concession is 2 months due for expiry, the email notification should be sent to both AE/AA and the BCM/Suite head.
When the concession is a month due for expiry, the email notification should be sent to the entire workflow, AE/AA, BCM[non-existing] and PCM or HO.
PCM and HO will never be never both be linked to a tblConcession record.
Use which ever one is available.